fix: resolve CI failures — clippy lints, lz4_flex advisory, cargo-deny
- Fix clippy errors: unnested or-patterns, zombie process, clone on Copy,
  length comparison, unnecessary cast, unused imports, doc indentation
- Allow clippy::unwrap_used in test/bench/example code (lib code is clean)
- Update lz4_flex 0.11.5 -> 0.11.6 (fixes RUSTSEC-2026-0041 vulnerability)
- Ignore RUSTSEC-2026-0002 (lru IterMut unsound) — transitive via ratatui,
  not used directly; will resolve when ratatui upgrades
- Fix unreachable_pub lint in test utility module

Refs PMAT-149

+++ examples/process_tracer_demo.rs
@@ -144,7 +144,7 @@ fn main() {
     // Create TraceResult with anomaly detection
     println!("\n7. Building TraceResult with anomaly detection...");
     let result =
-        TraceResult::new(std::process::id(), Duration::from_micros(total_time_us as u64), events)
+        TraceResult::new(std::process::id(), Duration::from_micros(total_time_us), events)
             .with_baseline(&baseline, config.anomaly_threshold);
 
     println!("   PID: {}", result.pid);
